Separate util imports
Importing all util classes had some detrimental problems because
the File/Analysis/ExperimentMetadata classes basically pulled in
all of Django, which can fail if certain environment variables aren't set.

This commit separates the imports so that the metadata classes aren't
automatically imported when importing any of the other util items.
