+++ scripts/data_loading/DCPEXPR0000000001_load_bounds_2021_scceres_experiment.py
@@ -12,7 +12,8 @@ from cegs_portal.search.models import (
     Facet,
     FacetValue,
 )
-from utils import ExperimentMetadata, timer
+from utils import timer
+from utils.experiment import ExperimentMetadata
 
 from . import get_closest_gene
