+++ scripts/data_loading/DCPEXPR0000000002_load_klann_2021_scceres_experiment.py
@@ -10,8 +10,9 @@ from cegs_portal.search.models import (
     DNAFeatureType,
     Experiment,
 )
-from utils import ExperimentMetadata, timer
+from utils import timer
 from utils.ccres import save_ccres
+from utils.experiment import ExperimentMetadata
 
 from . import get_closest_gene
