+++ scripts/data_loading/DCPEXPR0000000004_load_bounds_scCERES_mhc_2021_experiment.py
@@ -12,7 +12,8 @@ from cegs_portal.search.models import (
     Facet,
     FacetValue,
 )
-from utils import ExperimentMetadata, timer
+from utils import timer
+from utils.experiment import ExperimentMetadata
 
 from . import get_closest_gene
