+++ utils/__init__.py
@@ -1,4 +1,2 @@
-from .experiment import AnalysisMetadata, ExperimentMetadata
-from .file import FileMetadata
 from .misc import flatten, get_delimiter
 from .timing import timer
